Move `prepare_geometry` to `corrections` workflow
