+++ src/dolphin/workflows/corrections.py
@@ -1,14 +1,15 @@
 from __future__ import annotations
 
 import logging
+from collections.abc import Mapping, Sequence
 from dataclasses import dataclass
 from pathlib import Path
-from typing import Sequence
 
 from opera_utils import group_by_date
 
-from dolphin import PathOrStr, io, utils
+from dolphin import PathOrStr, io
 from dolphin._log import log_runtime, setup_logging
+from dolphin._types import Bbox
 from dolphin.atmosphere import estimate_ionospheric_delay
 from dolphin.timeseries import ReferencePoint
 from dolphin.workflows import CorrectionOptions
@@ -94,7 +95,7 @@ def run(
     epsg = crs.to_epsg()
     assert epsg is not None
     out_bounds = io.get_raster_bounds(timeseries_paths[0])
-    frame_geometry_files = utils.prepare_geometry(
+    frame_geometry_files = prepare_geometry(
         geometry_dir=geometry_dir,
         geo_files=correction_options.geometry_files,
         matching_file=Path(timeseries_paths[0]),
@@ -128,3 +129,118 @@ def run(
     )
 
     return CorrectionPaths(ionospheric_corrections=iono_paths)
+
+
+def prepare_geometry(
+    geometry_dir: Path,
+    geo_files: Sequence[Path],
+    matching_file: Path,
+    dem_file: Path | None,
+    epsg: int,
+    out_bounds: Bbox,
+    strides: Mapping[str, int] | None = None,
+) -> dict[str, Path]:
+    """Prepare geometry files.
+
+    Parameters
+    ----------
+    geometry_dir : Path
+        Output directory for geometry files.
+    geo_files : list[Path]
+        list of geometry files.
+    matching_file : Path
+        Matching file.
+    dem_file : Optional[Path]
+        DEM file.
+    epsg : int
+        EPSG code.
+    out_bounds : Bbox
+        Output bounds.
+    strides : Dict[str, int], optional
+        Strides for resampling, by default {"x": 1, "y": 1}.
+
+    Returns
+    -------
+    Dict[str, Path]
+        Dictionary of prepared geometry files.
+
+    """
+    from dolphin import stitching
+    from dolphin.io import DEFAULT_TIFF_OPTIONS, format_nc_filename
+
+    if strides is None:
+        strides = {"x": 1, "y": 1}
+    geometry_dir.mkdir(exist_ok=True)
+
+    stitched_geo_list = {}
+
+    if geo_files[0].name.endswith(".h5"):
+        # ISCE3 geocoded SLCs
+        datasets = ["los_east", "los_north", "layover_shadow_mask"]
+        nodatas = [0, 0, 127]
+
+        for nodata, ds_name in zip(nodatas, datasets):
+            outfile = geometry_dir / f"{ds_name}.tif"
+            logger.info(f"Creating {outfile}")
+            stitched_geo_list[ds_name] = outfile
+            ds_path = f"/data/{ds_name}"
+            cur_files = [format_nc_filename(f, ds_name=ds_path) for f in geo_files]
+
+            if ds_name not in "layover_shadow_mask":
+                options = (*DEFAULT_TIFF_OPTIONS, "NBITS=16")
+            else:
+                options = DEFAULT_TIFF_OPTIONS
+            stitching.merge_images(
+                cur_files,
+                outfile=outfile,
+                driver="GTiff",
+                out_bounds=out_bounds,
+                out_bounds_epsg=epsg,
+                in_nodata=nodata,
+                out_nodata=nodata,
+                target_aligned_pixels=True,
+                strides=strides,
+                resample_alg="nearest",
+                overwrite=False,
+                options=options,
+            )
+
+        if dem_file:
+            height_file = geometry_dir / "height.tif"
+            stitched_geo_list["height"] = height_file
+            if not height_file.exists():
+                logger.info(f"Creating {height_file}")
+                stitching.warp_to_match(
+                    input_file=dem_file,
+                    match_file=matching_file,
+                    output_file=height_file,
+                    resample_alg="cubic",
+                )
+    else:
+        # ISCE2 radar coordinates
+        dsets = {
+            "hgt.rdr": "height",
+            "incLocal.rdr": "incidence_angle",
+            "lat.rdr": "latitude",
+            "lon.rdr": "longitude",
+        }
+
+        for geo_file in geo_files:
+            if geo_file.stem in dsets:
+                out_name = dsets[geo_file.stem]
+            elif geo_file.name in dsets:
+                out_name = dsets[geo_file.name]
+                continue
+
+            out_file = geometry_dir / (out_name + ".tif")
+            stitched_geo_list[out_name] = out_file
+            logger.info(f"Creating {out_file}")
+
+            stitching.warp_to_match(
+                input_file=geo_file,
+                match_file=matching_file,
+                output_file=out_file,
+                resample_alg="cubic",
+            )
+
+    return stitched_geo_list
